Update moduleResolution in tsconfig
When trying to add new modules we were getting an issue where typescript
wasn't able to find them in the local node_modules folder, even though
they were there.

The reason is that the moduleResoluion was defaulting to 'classic' mode,
so in this PR we force it to 'node' which makes it check the node_modules
folder.